Rename more integration tests
